Convert e2e comparison benchmark to Vitest bench + CodSpeed
Rewrite the cross-library comparison (Sury vs Zod/Valibot/ArkType/TypeBox)
from Benchmark.js to Vitest `bench`, grouped by operation (create, parse,
create + parse, union) for head-to-head summaries. Wire the CodSpeed plugin
into the e2e Vitest config and run it in the Benchmarks CI job.

The old comparison.js targeted a removed API (S.parseOrThrow / S.compile) and
no longer ran; the rewrite uses the current S.parser(schema) compiled-fn API,
matching sury.bench.ts.
