+++ packages/e2e/package.json
@@ -9,7 +9,10 @@
     "ppx:install": "node ../sury-ppx/install.cjs",
     "ppx:build": "cd ../sury-ppx/src && dune build",
     "benchmark": "node ./src/benchmark/Benchmark.res.mjs",
-    "benchmark:comparison": "node ./src/benchmark/comparison.js"
+    "benchmark:comparison": "vitest bench --run"
+  },
+  "devDependencies": {
+    "@codspeed/vitest-plugin": "5.6.0"
   },
   "dependencies": {
     "rescript": "12.2.0",
